+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   ON-4044 # added `validateForm()` function for validating a form schema
+-   ON-4039 # added `generateFormElement()` function for generating & validating a form element
 
 ## 0.2.6 (2019-03-14)